backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
greenlet==3.0.3

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
aiohttp==3.9.1
httpx==0.25.2

# HTML parsing
beautifulsoup4==4.12.2

# Redis
redis==5.0.1

# Development and type checking
mypy==1.7.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
pre-commit==3.6.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Environment
python-dotenv==1.0.0